Buffer template writes
By writing to a buffer before writing to the HTTP response, we have the
chance to write HTTP errors instead. This makes it a lot easier to tell
when there's an error rendering a template. Before this, the template
would partially render and the error would get swallowed up, so the
client would just see a blank / half rendered page without any idea what
happened.